requirements.txt
boto==2.45.0
defusedxml==0.4.1
django==1.10.3
django-allauth==0.29.0
django-allauth-djrill==0.1.1
django-braces==1.10.0
django-countries-plus==1.1
django-model-utils==2.6
django-oauth-toolkit==0.10.0
django-querysetsequence==0.7
django-rest-auth==0.8.2
django-rest-framework-social-oauth2==1.0.4
django-webpack-loader==0.4.1
djangorestframework==3.5.3
djangorestframework-jwt==1.8.0
djrill==2.1.0
jsonfield==1.0.3
oauthlib==1.0.3
psycopg2==2.6.2
pycurl==7.43.0
PyJWT==1.4.2
python-dateutil==2.6.0
python-social-auth==0.2.21
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
urllib3==1.19
